refactor: make mailbox customizable

+++ bench/bench_test.go
@@ -42,7 +42,7 @@ func BenchmarkActor(b *testing.B) {
 		actorSystem, _ := actors.NewActorSystem("testSys",
 			actors.WithLogger(log.DiscardLogger),
 			actors.WithActorInitMaxRetries(1),
-			actors.WithMailboxSize(b.N),
+			actors.WithMailboxSize(uint64(b.N)),
 			actors.WithReplyTimeout(receivingTimeout))
 
 		// start the actor system
@@ -71,7 +71,7 @@ func BenchmarkActor(b *testing.B) {
 		actorSystem, _ := actors.NewActorSystem("testSys",
 			actors.WithLogger(log.DiscardLogger),
 			actors.WithActorInitMaxRetries(1),
-			actors.WithMailboxSize(b.N),
+			actors.WithMailboxSize(uint64(b.N)),
 			actors.WithExpireActorAfter(5*time.Second),
 			actors.WithReplyTimeout(receivingTimeout))
